1.2 проблемы с авторизацией

+++ index.php
@@ -14,6 +14,8 @@ $mail = new MailController($db, $user);
 $controller = new UserController($db);
 $id =  $user->getUserByLogin((string) $_SESSION['login'] ?? '');
 $form = $mail->get( (string) $id['id']);
+$auth = $user->issAuth();
+//var_dump($form);
 
 
 if (count($_POST)>0){
@@ -21,21 +23,21 @@ if (count($_POST)>0){
     if (isset($_POST['passwrod_confirmation'])){
         $controller->add();
     }
-    elseif (isset($_POST['login'])) {
+    elseif (isset($_POST['login']) && $auth) {
         $controller->in();
         header("Location: index.php");
         include_once 'view/form.php';
     }
-    elseif ((isset($_POST['out']))){
+    elseif ((isset($_POST['out'])) && $auth){
         $controller->out();
         header("Location: view/form.php");
     }
-    elseif (isset($_POST['send'])){
+    elseif (isset($_POST['send']) && $auth){
         $mail->send();
         header("Location: index.php");
         include_once 'view/form.php';
     }
-    elseif (isset($_POST['del'])){
+    elseif (isset($_POST['del']) && $auth){
         unset($_POST['del']);
         foreach ($_POST as  $value){
             $mail->delete($value);
@@ -54,7 +56,7 @@ if (count($_POST)>0){
     }
 }
 else {
-    if (!$user->issAuth()) {
+    if (!$auth) {
         include_once 'view/login.php';
     }
     else{
